Add sv_spree_players min for sprees (closed #8)
Do not track spree kills or deaths if player count is lower than cfg.
So when someone is on a spree x and then player count falls under the
threshold she wont loose her spree on death because kills do not count
either. A disconnect would end the spree tho. Otherwise the actual spree
tracking only continues when enough players are ingame again. Ingame not
online and spec or something.