Fallback directory for failed stats (closed #17)
If the writing of stats failed, for example due to a locked file, then
fallback to a second directory and save the account there.

Also added the admin command /merge_failed to then try to merge all the
failed stats back into the main directory and delete them from fallback
directory on success.